+++ Sources/Services/CustomerService.swift
@@ -163,7 +163,7 @@ extension CustomerService {
     func processCustomerAuthorizedEvent(_ event: CustomerAuthorizedEventDTO) throws {
         LogManager.trace("Processing customer authorized")
         
-        if connectionContext.accessToken != nil {
+        if connectionContext.channelConfig.isAuthorizationEnabled {
             guard let token = event.postback.data.accessToken else {
                 throw CXoneChatError.missingAccessToken
             }